+++ changelog.md
@@ -1,2 +1,2 @@
 ### Trash Cans 1.0.17
-- Initial release of Trash Cans for Fabric
+- Initial release of Trash Cans for Minecraft 1.20